# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKF Racing Hub API",
    "description": "REST API of the SKF Racing Hub site.\n\n**Authentication.** A `session_id` cookie, set by the Discord login. Operations\nthat need a session list it under *security*; the roles they accept are in\n`x-required-roles`.\n\n**Errors.** Every error is an `application/problem+json` document (`Problem`).\n`code` is a stable machine-readable identifier, `detail` a message that can be\nshown to the user. Validation failures (422) also carry `errors`, one entry per\nrejected field.\n\n**Paging.** Collections that take `limit` and `offset` return a plain array;\nthe total is in `X-Total-Count` and neighbouring pages in an RFC 8288 `Link`\nheader.\n\n**Tracing.** Every response carries `X-Request-ID`; quote it when reporting a\nproblem.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/championships/{championshipId}/races": {
      "get": {
        "tags": [
          "Championships"
        ],
        "summary": "List championship races",
        "description": "All races of the championship, including future ones.",
        "operationId": "listChampionshipRaces",
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChampionshipRaceOut"
                  }
                }
              }
            },
            "headers": {
              "X-Data-Stale": {
                "description": "`true` when SimGrid was unreachable and cached data was served.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/api/v1/championships/{championshipId}/races/{raceId}/results": {
      "get": {
        "tags": [
          "Championships"
        ],
        "summary": "Get race results",
        "description": "One race's classification (race or qualifying), ordered by class.",
        "operationId": "getRaceResults",
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RaceSessionKind",
              "default": "race"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceSessionOut"
                }
              }
            },
            "headers": {
              "X-Data-Stale": {
                "description": "`true` when SimGrid was unreachable and cached data was served.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
...
    "/api/v1/custom-championships/{championshipId}/races": {
      "post": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Create custom race",
        "operationId": "createCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRaceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRaceOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "put": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Replace custom races",
        "description": "Replace the full race list in one request.\n\nRequires role: `admin`, `super_admin`, `community_manager`.",
        "operationId": "replaceCustomRaces",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustomRaceUpsert"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomRaceOut"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ]
      }
    },
    "/api/v1/custom-championships/{championshipId}/races/{raceId}": {
      "get": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Get custom race",
        "operationId": "getCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRaceOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "patch": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Update custom race",
        "operationId": "updateCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomRaceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomRaceOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      },
      "delete": {
        "tags": [
          "Custom championships"
        ],
        "summary": "Delete custom race",
        "operationId": "deleteCustomRace",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "raceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "championshipId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin",
          "community_manager"
        ],
        "description": "Requires role: `admin`, `super_admin`, `community_manager`."
      }
    },
...
    "/api/v1/race-result-imports": {
      "get": {
        "tags": [
          "Race results"
        ],
        "summary": "List race result imports",
        "operationId": "listRaceResultImports",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "championshipId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RaceResultImportOut"
                  }
                }
              }
            },
            "headers": {
              "X-Total-Count": {
                "description": "Size of the whole collection, across all pages.",
                "schema": {
                  "type": "integer"
                }
              },
              "Link": {
                "description": "RFC 8288 links to the `next` and `prev` pages, when they exist.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "post": {
        "tags": [
          "Race results"
        ],
        "summary": "Create race result import",
        "description": "Import one round's result file, replacing any earlier upload of it.\n\nRequires role: `admin`, `super_admin`.",
        "operationId": "createRaceResultImport",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RaceResultImportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceResultImportResultOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/ContentTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ]
      }
    },
    "/api/v1/race-result-imports/{importId}": {
      "get": {
        "tags": [
          "Race results"
        ],
        "summary": "Get race result import",
        "operationId": "getRaceResultImport",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceResultImportOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "delete": {
        "tags": [
          "Race results"
        ],
        "summary": "Delete race result import",
        "description": "Remove the round's results. Its incidents stay with the window.\n\nRequires role: `admin`, `super_admin`.",
        "operationId": "deleteRaceResultImport",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ]
      }
    },
    "/api/v1/race-result-imports/{importId}/file": {
      "get": {
        "tags": [
          "Race results"
        ],
        "summary": "Download race result import file",
        "operationId": "downloadRaceResultImportFile",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The file as it was uploaded.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "description": "`attachment` with the original file name.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
    "/api/v1/race-result-imports/{importId}/parse-runs": {
      "post": {
        "tags": [
          "Race results"
        ],
        "summary": "Create parse run",
        "description": "Re-run the parser on the stored original, e.g. after a parser fix.\n\nA parse run replaces the import with a new one (new id), so this answers\n201 with the replacement's ``Location``.\n\nRequires role: `admin`, `super_admin`.",
        "operationId": "createParseRun",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceResultImportResultOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionshipRaceOut": {
        "properties": {
          "id": {
            "type": "integer"
          },
          "displayName": {
            "type": "string",
            "default": ""
          },
          "startsAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "track": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "resultsAvailable": {
            "type": "boolean",
            "default": false
          },
          "ended": {
            "type": "boolean",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "displayName",
          "startsAt",
          "track",
          "resultsAvailable",
          "ended"
        ]
      },
...
      "CustomRaceCreate": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "track": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object"
      },
      "CustomRaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "track": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "sortOrder": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "endDate",
          "track",
          "sortOrder",
          "createdAt"
        ]
      },
...
      "CustomRaceUpsert": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "track": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "description": "Used for batch race sync — id present means update, absent means create."
      },
...
      "RaceSessionKind": {
        "type": "string",
        "enum": [
          "race",
          "qualifying"
        ]
      },
      "RaceSessionOut": {
        "properties": {
          "raceId": {
            "type": "integer"
          },
          "session": {
            "$ref": "#/components/schemas/RaceSessionKind"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/RaceResultEntryOut"
            },
            "type": "array",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "raceId",
          "session",
          "entries"
        ]
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session id, set as a cookie by the Discord login.",
        "in": "cookie",
        "name": "session_id"
      },
      "ingestToken": {
        "type": "http",
        "description": "Shared token of the external incident parsers.",
        "scheme": "bearer"
      }
    },
    "responses": {
      "BadGateway": {
        "description": "Bad Gateway",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "UnprocessableContent": {
        "description": "Unprocessable Content",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblem"
            }
          }
        }
      }
    }
  },
...
}